chore(types): collect types in top-level types.ts
