+++ src/types.ts
@@ -3,11 +3,11 @@
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface JSONArray extends Array<JSONValue> {}
 type JSONValue = string | number | boolean | JSONObject | JSONArray
-interface JSONObject {
+export interface JSONObject {
   [x: string]: JSONValue
 }
-type BaseConfig = JSONObject
-type Schema = JSONObject
+export type BaseConfig = JSONObject
+export type Schema = JSONObject
 
 type SopsMetadata = {
   kms: unknown | null
@@ -19,7 +19,18 @@ type SopsMetadata = {
   version: string
   encrypted_suffix?: string | null
 }
-type EncryptedConfig = { sops?: SopsMetadata } & BaseConfig
-type DecryptedConfig = Omit<BaseConfig, 'sops'>
-type HydratedConfig = { runtimeEnvironment: string } & DecryptedConfig
-type MemoizedConfig = HydratedConfig | undefined
+
+export type EncryptedConfig = { sops?: SopsMetadata } & BaseConfig
+export type DecryptedConfig = Omit<BaseConfig, 'sops'>
+export type HydratedConfig = { runtimeEnvironment: string } & DecryptedConfig
+export type MemoizedConfig = HydratedConfig | undefined
+
+export enum FileExtension {
+  JSON = 'json',
+  YAML = 'yaml',
+  YML = 'yml',
+}
+export type File = {
+  contents: EncryptedConfig | Schema
+  filePath: string
+}
